ENH: Wrap VectorImage with complex real pixel types
VariableLengthVector also needs to be wrapped as a dependency.

This supports wrapping in the IsotropicWavelet module.

double cast is remove to address the error:

/home/matt/src/ITK3/Modules/Core/Common/include/itkVariableLengthVector.hxx:449:43: error: cannot convert 'RealValueType' (aka 'complex<double>') to 'double' without a conversion operator
  return static_cast<RealValueType>(sqrt( double( this->GetSquaredNorm() ) ) );
                                          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
/home/matt/bin/ITK3-Wrap-MinSizeRel/Wrapping/Modules/ITKCommon/itkVariableLengthVectorPython.cpp:6053:59: note: in instantiation of member function 'itk::VariableLengthVector<std::complex<float> >::GetNorm' requested here
      result = ((itkVariableLengthVectorCF const *)arg1)->GetNorm();

+++ Modules/Core/Common/include/itkVariableLengthVector.hxx
@@ -446,7 +446,7 @@ VariableLengthVector< TValue >
 ::GetNorm(void) const ITK_NOEXCEPT
 {
   using std::sqrt;
-  return static_cast<RealValueType>(sqrt( double( this->GetSquaredNorm() ) ) );
+  return static_cast<RealValueType>(sqrt( this->GetSquaredNorm() ) );
 }
 
 /**
